pyproject.toml
[tool.poetry]
name = "arya_date"
version = "0.1.0"
description = ""
authors = ["adimyth <mishraaditya6991@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.1.3"
matplotlib = "^3.3.2"
sklearn = "^0.0"
black = "^20.8b1"
isort = "^5.6.4"
opencv-python = "^4.4.0"
tqdm = "^4.51.0"
pydot = "^1.4.1"
graphviz = "^0.14.2"
pillow = "^8.0.1"
scikit-image = "^0.17.2"
tensorflow = "2.3.0"
jupyterlab = "^2.2.9"
rapids = "^0.0.1"
efficientnet = "^1.1.1"
iterative-stratification = "^0.1.6"
jupyter_contrib_nbextensions = "^0.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
